Temporary Fix with Concretization and Invalidation of Expr's users.
Once ewe concretized Expr, we should be sure that our users are invalidated, otherwise, we might be in situtation when Expr is invalidated, users partially concretized, but users of the users are not. And cached invalid z3 expr gonna be emitted to a solver. Currently, we just blatantly try to invalidate a tree of users.